% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_bounds.R
\name{censorThis}
\alias{censorThis}
\alias{censorThis,WoodenHorse,ANY-method}
\alias{censorThis,numeric,numeric_OR_missing-method}
\title{Censor any values outside of range}
\usage{
censorThis(x, range = NULL, only.finite = TRUE, aes = "z")

\S4method{censorThis}{WoodenHorse,ANY}(x, range = c(0, 1), only.finite = TRUE, aes = "z")

\S4method{censorThis}{numeric,numeric_OR_missing}(x, range = c(0, 1), only.finite = TRUE, aes = "z")
}
\arguments{
\item{x}{A vector of values to manipulate.}

\item{range}{A range representing object.}

\item{only.finite}{If \code{TRUE} (the default), will only modify finite
values.}

\item{aes}{An aesthetic for which to evaluate the function.}
}
\value{
The \code{x} argument, but with an indication which values should be
  censored.
}
\description{
This function follows the \code{\link[scales]{censor}} function,
  but allows extensions that work with objects other than numeric vectors.
}
\examples{
# For regular numeric vectors
censorThis(c(-1, 0.5, 1, 2, NA))
}
