% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicmanip.R
\name{check_valid_hiclayer}
\alias{check_valid_hiclayer}
\title{Checks for validity of a Hi-C layer}
\usage{
check_valid_hiclayer(exp1, exp2, xranges, yranges)
}
\arguments{
\item{exp1, exp2}{A GENOVA experiment object.}

\item{xranges, yranges}{A GRanges object. Alternatively,a \code{list} of at least length 3 with the
following elements: \describe{\item{\code{character}}{ vector of chromosome
names}\item{\code{integer}}{ vector of start
positions}\item{\code{integer}}{ vector of end positions}}}
}
\value{
A \code{logical} of length 1.
}
\description{
Checks for validity of a Hi-C layer
}
\keyword{internal}
