% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideograms.R
\name{clear_cytoband_cache}
\alias{clear_cytoband_cache}
\title{Clear the cytoband cache.}
\usage{
clear_cytoband_cache()
}
\description{
The ggnomics package creates a small environment to store ideogram related
functions and data. When switching organisms in a single session, it might be
useful to clear the ideogram data from the previous organism to make room
for the current organism.
}
\examples{
clear_cytoband_cache()
}
