% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_constructors.R
\name{coord_S4}
\alias{coord_S4}
\title{Cartesian coordinates for S4 classes}
\usage{
coord_S4(xlim = NULL, ylim = NULL, expand = TRUE, default = FALSE, clip = "on")
}
\arguments{
\item{xlim}{Limits for the x and y axes.}

\item{ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{default}{Is this the default coordinate system? If \code{FALSE} (the default),
then replacing this coordinate system with another one creates a message alerting
the user that the coordinate system is being replaced. If \code{TRUE}, that warning
is suppressed.}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of \code{"on"} (the default) means yes, and a setting of \code{"off"}
means no. In most cases, the default of \code{"on"} should not be changed,
as setting \code{clip = "off"} can cause unexpected results. It allows
drawing of data points anywhere on the plot, including in the plot margins. If
limits are set via \code{xlim} and \code{ylim} and some data points fall outside those
limits, then those data points may show up in places such as the axes, the
legend, the plot title, or the plot margins.}
}
\description{
The Cartesian coordinate system is the typical linear coordinate
  system wherein a point on the plane is parameterised by the x- and
  y-position and distances between point are Euclidean. The S4 variant allows
  the contextual evaluation of, for example, ranged data classes.
}
\details{
This coordinate function is designed to work with S4 scales. In
  absence thereof, no contextual evaluation should occur.

  As a temporary workaround, when the coordinates are set up, all S4 Vector
  classes are disguised as the \code{WoodenHorse} class.
}
\examples{
NULL
}
