% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideograms.R
\name{draw_ideo_panels_grid}
\alias{draw_ideo_panels_grid}
\title{Drawing panels for facet_ideogrid()}
\usage{
draw_ideo_panels_grid(
  panels,
  layout,
  x_scales,
  y_scales,
  ranges,
  coord,
  data,
  theme,
  params,
  self
)
}
\value{
A gtable object
}
\description{
This function handles the assembly of panels into a
\code{\link[gtable]{gtable}} object. This is a method of the "FacetIdeoGrid"
\code{\link[ggplot2]{ggproto}} class. This method recieves from objects
elsewhere in the plot a list of \code{\link[grid]{grid.grob}} with the
contents of panels. This method is responsible for decorating the panels with
axes, strips and ideograms and positioning these together in a gtable.
}
\details{
Since this is a method of a class, this function should only be used
  indirectly by users.
}
\keyword{internal}
