% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-format.R
\name{est_accuracy}
\alias{est_accuracy}
\title{Estimating the needed accuracy
Mainly to get around r-lib/scales#251 issue}
\usage{
est_accuracy(x)
}
\description{
Estimating the needed accuracy
Mainly to get around r-lib/scales#251 issue
}
