% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_cytoband_colours}
\alias{example_cytoband_colours}
\title{Example colours for cytobands.}
\usage{
example_cytoband_colours()
}
\value{
A character vector of colours named by \code{gieStain} levels.
}
\description{
Gives a selection of colours compatible with the \code{gieStain} column of
\code{example_cytobands}.
}
\examples{
cytocols <- example_cytoband_colours()
}
\seealso{
\code{\link[ggnomics]{example_cytobands}}
}
