% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_HiC}
\alias{example_HiC}
\title{Generate fake Hi-C experiment}
\usage{
example_HiC()
}
\value{
A \code{list} with the skeleton of a GENOVA Hi-C experiment.
}
\description{
Randomly generate data in the format of a GENOVA Hi-C experiment.
}
\details{
The data is formatted as the first 120 Mb of 'chr1' at 1 Mb
  resolution. Since data is randomly generated, use \code{set.seed()} for
  reproducibility. Qualifies as the bare minimum requirement for the
  \code{\link[ggnomics]{geom_hictriangle}}'s \code{exp} argument.
}
\examples{
exp <- example_HiC()
}
