% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicmanip.R
\name{extract_hicdata}
\alias{extract_hicdata}
\title{Extractor function for Hi-C GENOVA experiment objects.}
\usage{
extract_hicdata(exp1, exp2 = NULL, xranges, yranges = NULL, triangle = FALSE)
}
\arguments{
\item{exp1, exp2}{A GENOVA experiment object.}

\item{xranges, yranges}{A GRanges object. Alternatively,a \code{list} of at
least length 3 with the following elements:
\describe{\item{\code{character}}{ vector of chromosome
names}\item{\code{integer}}{ vector of start
positions}\item{\code{integer}}{ vector of end positions}}}

\item{triangle}{A \code{logical} of length 1. If \code{TRUE}, only return
half a Hi-C matrix.}
}
\value{
A formatted \code{data.frame}
}
\description{
Extractor function for Hi-C GENOVA experiment objects.
}
\keyword{internal}
