% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_ideogrid.R
\name{facet_ideogrid}
\alias{facet_ideogrid}
\title{Lay out panels in a grid with ideograms}
\usage{
facet_ideogrid(rows = NULL, cols = NULL, scales = "fixed",
  space = "fixed", shrink = TRUE, labeller = "label_value",
  as.table = TRUE, switch = NULL, drop = TRUE, margins = FALSE,
  facets = NULL, ideo.size = unit(0.1, "null"), high.col = NA)
}
\arguments{
\item{rows, cols, scales, space, shrink, labeller, as.table, switch, drop, margins, facets}{see \code{\link[ggplot2]{facet_grid}}}

\item{ideo.size}{\code{\link[grid]{unit}} value object of class \code{"unit"}
specifying ideogram size.}

\item{high.col}{\code{NA} (default) or colour value of ideogram highlights
(see details). If \code{NA}, no highlighting will occur.}
}
\description{
\code{facet_ideogrid()} mimicks behaviour of
\code{\link[ggplot2]{facet_grid}} but adds ideograms. The facetting variables
are checked for chromosome names and if these are found, then ideograms are
plotted in between the strips and panels.
}
\details{
In absence of chromosome names in variables will result in default
  \code{facet_grid()} behaviour. A mix of chromosome names and other
  variables results in whitespace between the panels and other variable
  strips.

  By default, this function will not highlight the panel axis range. To
  highlight the plotted region on the ideogram, set \code{high.col} to a
  colour. If highlights occur, region to be highlighted will be computed from the
  relevant axis limits of the corresponding panels.

  Colours can be specified in one of the forms returned by
  \code{\link[grDevices]{rgb}}, as name (see \code{\link[grDevices]{colors}})
  or as non-negative integer index into the current
  \code{\link[grDevices]{palette}} (with zero being taken as transparent).
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) +
   geom_point() +
   facet_ideogrid(~ "chr1")
}
\seealso{
\code{\link[ggplot2]{facet_grid}} \code{\link[grid]{unit}}
}
