% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_track.R
\name{facet_track}
\alias{facet_track}
\title{Lay out panels as tracks}
\usage{
facet_track(heights = NULL, widths = NULL, respect = FALSE,
  scales = "free", space = "fixed", shrink = TRUE,
  labeller = "label_value", as.table = TRUE, switch = "y",
  drop = TRUE, margins = FALSE)
}
\arguments{
\item{heights}{a \code{numeric} or \code{unit} vector for setting panel
heights.}

\item{widths}{a \code{numeric} or \code{unit} vector for setting panel
widths.}

\item{respect}{a \code{logical} value. If \code{TRUE}, widths and heights
specified in \code{"null"} units are proportional. If \code{FALSE},
\code{"null"} units in the x- and y-direction vary independently.}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with formulae of the type \code{~cyl + am}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[=labeller]{labeller()}}. See
\code{\link[=label_value]{label_value()}} for more details and pointers to other
options.}

\item{as.table}{If \code{TRUE}, the default, the facets are laid out like
a table with highest values at the bottom-right. If \code{FALSE}, the
facets are laid out like a plot with the highest value at the top-right.}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}

\item{margins}{Either a logical value or a character
vector. Margins are additional facets which contain all the data
for each of the possible values of the faceting variables. If
\code{FALSE}, no additional facets are included (the
default). If \code{TRUE}, margins are included for all faceting
variables. If specified as a character vector, it is the names of
variables for which margins are to be created.}
}
\description{
\code{facet_track} acts much like \code{facet_grid}, but looks for fixed
facetting variables in each layer's data. Furthermore, allows panel
dimensions can be set.
}
\details{
The primary use case for \code{facet_track} is to be used in
  combination with the \code{\link[ggnomics]{tracklayer}} function and
  allowing panel specification to be independent of the \code{data} arguments
  of geoms and stats. \code{facet_track} looks in the layer data for the
  columns named \code{"track_x"} and \code{"track_y"} to evaluate facet
  specification.

  When \code{heights} or \code{widths} are \code{numeric} vectors, panel
  sizes are defined as ratios i.e. relative \code{"null"} units.
  \code{heights} and \code{widths} vectors are repeated or shortened to fit
  the number of panels in their directions. When \code{heights} or
  \code{widths} are \code{NULL}, no changes from the default settings are
  made in that directions. This can be convenient when '\code{space}' is set
  to \code{"free_*"}. Setting the sizes of panels overrides any aspect ratios
  set in the theme and \code{coord} functions.
}
\examples{
ggplot() +
  tracklayer("point", data = mtcars,
             mapping = aes(disp, mpg),
             track_x = "mtcars") +
  tracklayer("point", data = iris,
             mapping = aes(Sepal.Width, Sepal.Length),
             track_x = "iris") +
  facet_track(widths = c(2, 1))
}
