% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperscales.R
\name{format_genco}
\alias{format_genco}
\title{Formatter for genomic coordinates}
\usage{
format_genco(x)
}
\arguments{
\item{x}{A \code{numeric} vector to format}
}
\value{
A \code{character} vector of formatted values
}
\description{
Formats values from \code{"1e+06"} notation to \code{"1 Mb"}.
}
\details{
Written as a convenience function to make genomic coordinates more
  pleasing to see.
}
\examples{
format_genco(10^seq(0, 7, by = 1))
}
