% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperscales.R
\name{format_logtrans}
\alias{format_logtrans}
\title{Formatter for log-transformed values}
\usage{
format_logtrans(x)
}
\arguments{
\item{x}{A \code{numeric} vector to format.}
}
\value{
A list of calls for formatted expressions.

Written as a convenience funtion to make scales of log-transformed plots easier on the eyes.
}
\description{
Formats values from 1e+06 notation to 10^6 notation.
}
\examples{
y <- format_logtrans(10^c(1:3))
}
