% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperscales.R
\name{format_megabase}
\alias{format_megabase}
\title{Formatter for megabasepairs}
\usage{
format_megabase(x)
}
\arguments{
\item{x}{A \code{numeric} vector to format.}
}
\value{
A character vector of formatted values.
}
\description{
Formats values from "1e+06" notation to "1 Mb"
}
\details{
Written as convenience function to make chromosome scales easier to read.
}
\examples{
format_megabase(c(1:3)*1e7)
}
