% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hictriangle.R
\docType{data}
\name{geom_hictriangle}
\alias{geom_hictriangle}
\alias{GeomHicTriangle}
\title{Hi-C Triangle}
\format{An object of class \code{GeomHicTriangle} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
geom_hictriangle(exp, ranges, stat = "identity", position = "identity",
  ..., na.rm = FALSE, show.legend = NA)

GeomHicTriangle
}
\arguments{
\item{exp}{A GENOVA experiment}

\item{ranges}{A GRanges object. Alternatively, a \code{list} of at least
length 3 with the following elements: \describe{\item{\code{character}}{
vector of chromosome names}\item{\code{integer}}{ vector of start
positions}\item{\code{integer}}{ vector of end positions}}}

\item{stat, position, ..., na.rm, show.legend, inherit}{See
\code{\link[ggplot2]{geom_polygon}}}
}
\description{
The Hi-C triangle geom is used to portray a rotated Hi-C matrix such that the
diagonal runs along the x-axis. It can be convenient for displaying details
located in a region of interest and proximal to the diagonal.
}
\details{
The triangle is calculated such that it is rotated by 45 degrees
  clockwise and rescaled such that the x-axis is equivalent to the original
  region and the y-axis corresponds to distance from the diagonal.
  \code{geom_hictriangle} invokes these steps at the layer level by
  specifying how a raster-like Hi-C matrix should be setup as a triangle.

  A recommendation for this geom is to use a \code{coord_fixed(ratio = 0.5)}
  to get neat 45 degree angles.
}
\seealso{
\code{\link[GENOVA]{construct.experiment}} on how to construct a
  GENOVA experiment object. \code{\link[GenomicRanges]{GRanges}} on how to
  make or subset GRanges objects.
}
\keyword{datasets}
