% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pointpath.R
\docType{data}
\name{geom_pointpath}
\alias{geom_pointpath}
\alias{GeomPointPath}
\title{Point Paths}
\usage{
geom_pointpath(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
\code{\link[ggplot2]{geom_point}}}
}
\description{
The point path geom is used to make a scatterplot wherein the order of points
if of interest. This geom was created to mimick the \code{type = 'b'} style
of base R line plots.
}
\details{
The \code{linesize} aesthetic can be interpreted as the \code{size}
  aesthetic for \code{geom_line()}. The \code{mult} is a numeric value to
  scale the proportion of gaps in the line around points.
}
\section{Aesthetics}{
 \code{geom_pointpath()} understands the following
  aesthetics (required aesthetics are in bold):
  \itemize{\item{\strong{\code{x}}} \item{\strong{\code{y}}}
  \item{\code{alpha}} \item{\code{colour}} \item{\code{group}}
  \item{\code{shape}} \item{\code{size}} \item{\code{stroke}}
  \item{\code{linesize}} \item{\code{linetype}} \item{\code{mult}}}
}

\examples{
ggplot(pressure, aes(temperature, pressure)) +
  geom_pointpath()
}
\keyword{datasets}
