% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_range.R
\name{geom_range}
\alias{geom_range}
\title{Ranges}
\usage{
geom_range(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \code{Layer} ggproto object.
}
\description{
Alternative parameterisation of rectangles which can take a
\code{Ranges}-class, such as \linkS4class{IRanges} or \linkS4class{GRanges}
as x or y argument.
}
\details{
If the \code{x} position aesthetic is given while \code{y} is not,
and vice versa, the default behaviour is to center the aesthetic that is not
provided around zero with a height/width of 0.8. If the position aesthetic is
\code{numeric}, it is expected that also the \code{(x|y)end} aesthetic is
provided.
}
\section{Aesthetics}{
 \code{geom_range} understands the following
aesthetics (required aesthetics are in bold, optional in italic).
\itemize{
 \item \strong{x}
 \item \strong{y}
 \item \emph{xend} (if data is numeric)
 \item \emph{yend} (if data is numeric)
 \item alpha
 \item colour
 \item fill
 \item group
 \item linetype
 \item size
 \item weight
}
}

\examples{
require(IRanges)
df <- DataFrame(x = IRanges(c(1, 2), c(3, 4)))

ggplot(df) +
  geom_range(aes(x))
}
