% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnomics-package.R
\docType{package}
\name{ggnomics-package}
\alias{ggnomics}
\alias{ggnomics-package}
\title{ggnomics: ggnomics}
\description{
Provides tools for the integration of Bioconductor S4 classes and 
    the popular 'ggplot2' plotting package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teunbrand/ggnomics}
  \item Report bugs at \url{https://github.com/teunbrand/ggnomics/issues}
}

}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
