% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_nested.R, R/facet_track.R,
%   R/geom_genemodel.R, R/geom_hictriangle.R, R/geom_pointpath.R,
%   R/geom_polygonraster.R, R/geom_rectrug.R, R/ggnomics_extensions.R,
%   R/layer_track.R, R/position_disjoint_ranges.R, R/position_lineartrans.R,
%   R/stat_theodensity.R
\docType{data}
\name{FacetNested}
\alias{FacetNested}
\alias{FacetTrack}
\alias{GeomGeneModel}
\alias{GeomHicTriangle}
\alias{GeomPointPath}
\alias{GeomPolygonRaster}
\alias{GeomRectMargin}
\alias{GeomTileMargin}
\alias{ggnomics_extensions}
\alias{TrackLayer}
\alias{PositionDisjointRanges}
\alias{PositionLinearTrans}
\alias{StatTheoDensity}
\title{ggnomics extensions to ggplot2}
\description{
ggnomics relies on the extension mechanism of ggplot2 through ggproto class
objects, which allows cross-package inheritance of objects such as geoms,
stats, facets, scales and coordinate systems. These objects can be ignored by
users for the purpose of making plots, since interacting with these objects
is preferred through various geom_\*, stat_\*, facet_\*, coord_\* and scale_\*
functions.
}
\seealso{
\link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
