% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto_sibling.R
\name{ggproto_sibling}
\alias{ggproto_sibling}
\title{Like 'ggproto', but also captures methods from 'sister' in addition to the
usual parental inheritance. Used in the ScaleS4-family of scale ggprotos.}
\usage{
ggproto_sibling(class_name = character(), parent = NULL, sister = NULL, ...)
}
\description{
Like 'ggproto', but also captures methods from 'sister' in addition to the
usual parental inheritance. Used in the ScaleS4-family of scale ggprotos.
}
