% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_axis_genomic.R
\name{guide_genomic_axis}
\alias{guide_genomic_axis}
\alias{guide_train.genomic_axis}
\alias{guide_transform.genomic_axis}
\alias{guide_gengrob.genomic_axis}
\title{Axis for genomic positions}
\usage{
guide_genomic_axis(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver()
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in \code{\link[ggplot2:theme]{theme()}} / \code{\link[ggplot2:element_text]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\value{
A \code{guide} object.
}
\description{
This axis guide is the genomic equivalent of
\code{\link[ggplot2]{guide_axis}} to pair with the genomic position scales
\code{\link[=scale_genomic]{scale_(x|y)_genomic()}}.
}
\details{
This guide places the sequence names in the middle or their limits
  and places label for minor breaks at positions along the sequences.
  Defaults to \code{guide_axis} behaviour if not exposed to genomic data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{guide_train}: Trainer for genomic axis. See
\code{\link[ggplot2]{guide-exts}}.

\item \code{guide_transform}: Transformer for genomic axis. See
\code{\link[ggplot2]{guide-exts}}.

\item \code{guide_gengrob}: Graphic object generator for genomic axis. See
\code{\link[ggplot2]{guide-exts}}.
}}

\examples{
NULL
}
