% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_axis_ideogram.R
\name{guide_ideogram_axis}
\alias{guide_ideogram_axis}
\alias{guide_train.ideogram_axis}
\alias{guide_transform.ideogram_axis}
\alias{guide_gengrob.ideogram_axis}
\title{Ideogram axis guide}
\usage{
guide_ideogram_axis(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver(),
  chromosome = NULL,
  genome = NULL,
  ideo_size = unit(0.5, "cm"),
  funnel_size = unit(0.5, "cm"),
  funnel_col = "dodgerblue",
  regular_axis = TRUE
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in \code{\link[ggplot2:theme]{theme()}} / \code{\link[ggplot2:element]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{chromosome}{A \code{character} with chromosome name of length one, for
example \code{"chr12"}.}

\item{genome}{A \code{character} with a genome name of length one, for
example \code{"hg38"}.}

\item{ideo_size}{A \code{unit} of length one indicating the size of the
ideogram. If a \code{numeric} is provided, this is interpreted as
millimetre units. A small proportional margin will be added for an
indicator.}

\item{funnel_size}{A \code{unit} of length one indicating the size of the
funnel space. If a \code{numeric} is provided, this is interpreted as
millimetre units. If this can be interpreted as zero, no funnel will be
drawn.}

\item{funnel_col}{A \code{character} of length one setting the colour for the
funnel and indicator. Set to \code{NA} to make the funnel and indicator
transparent.}

\item{regular_axis}{A \code{logical} of length one: should the ticks and
labels of the usual position axis be drawn? (default: \code{TRUE}).}
}
\value{
A list with the S3 class \code{guide}.
}
\description{
Places an ideogram next to the axis and indicates the data range interpreted
as genomic positions on the ideogram. The ideograms need to be setup first
with the \code{setup_ideogram()} function.
}
\details{
Currently does not support S4Vector or genomic scales (yet).
}
\section{Methods (by generic)}{
\itemize{
\item \code{guide_train}: Trainer for ideogram axis. See
\code{\link[ggplot2]{guide-exts}}.

\item \code{guide_transform}: Transformer for ideogram axis. See
\code{\link[ggplot2]{guide-exts}}.

\item \code{guide_gengrob}: Graphic object generator for ideogram axis.
See \code{\link[ggplot2]{guide-exts}}.
}}

\examples{
# Setup ideograms and make a basic plot
setup_ideograms(genome = "hg38")
p <- ggplot(iris, aes(Sepal.Width * 1e7, Sepal.Length * 1e7)) +
    geom_point()

# As main axis
p + guides(x = guide_ideogram_axis("hg38", chromosome = "chr2"))

# As secundary axis
p + guides(x.sec = guide_ideogram_axis("hg38", chromosome = "chr2",
                                       regular_axis = FALSE))

# Without funnel
p + guides(x = guide_ideogram_axis("hg38", chromosome = "chr2",
                                   funnel_size = 0))

# Without funnel or indicator
p + guides(x = guide_ideogram_axis("hg38", chromosome = "chr2",
                                   funnel_size = 0, funnel_col = NA))
}
\seealso{
The \code{\link[ggnomics]{setup_ideograms}()} function for importing
  and caching ideogram data.

Other position guides: 
\code{\link{guide_genomic_axis}()}
}
\concept{position guides}
