% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_extract.R
\name{hic_extractor}
\alias{hic_extractor}
\title{Extract Hi-C data from GENOVA contacts objects.}
\usage{
hic_extractor(exp1, exp2 = NULL, xrange = NULL, yrange = NULL)
}
\arguments{
\item{exp1, exp2}{A GENOVA \code{contacts} object.}

\item{xrange, yrange}{A \code{GRanges} object. Alternatively, a three-column
\code{data.frame} with the following column structure (column names not required):
\describe{
  \item{chromosome}{a character vector noting the chromosome name}
  \item{start}{an integer vector indicating the start position of the genomic range}
  \item{end}{an integer vector indicating the end position of the genomic range}
}}
}
\value{
A \code{data.frame} with the following columns:
  \describe{
    \item{\code{xidx}}{The Hi-C bin index for the x-position}
    \item{\code{yinx}}{The Hi-C bin index for the y-position}
    \item{\code{contacts}}{A numerical value representing the contacts between \code{xidx} and \code{yidx}}
    \item{\code{entry}}{An index into the '\code{xrange}' argument.}
    \item{\code{exp}}{An index determining the origin of the data, either 1 for the '\code{exp1}' argument or 2 for the '\code{exp2}' argument.}
    \item{\code{x}}{The basepair position of the start of \code{xidx}}
    \item{\code{y}}{The basepair position of the start of \code{yidx}}
    \item{\code{xchr}}{The chromosome for \code{xidx}}
    \item{\code{ychr}}{The chromosome for \code{yidx}}
  }
}
\description{
A convenience function for extracting 3D genome contact information from
\pkg{GENOVA}'s \code{contacts} objects.
}
\details{
When the '\code{exp2}' argument is not \code{NULL}, the second
  experiment must be compatible with the first in having identical indices.

  The '\code{yrange}' argument must either be \code{NULL}, describe a single
  genomic range, or an equal number of genomic ranges as the '\code{xrange}'
  argument. Furthermore, the '\code{yrange}' argument must be strictly
  identical or downstream to the '\code{xrange}' argument.
}
\examples{
x <- 1:10
}
\seealso{
For the \code{contacts} constructor: \code{\link[GENOVA]{load_contacts}}.
}
