% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideograms.R
\name{render.ideo}
\alias{render.ideo}
\title{Ideogram renderer}
\usage{
render.ideo(chr, orient = "x", ranges, theme, high.col, ideodat)
}
\arguments{
\item{chr}{a chromosome name.}

\item{orient}{a character "x" or "y".}

\item{ranges}{the limits of the panel.}

\item{theme}{the theme of the plot.}

\item{high.col}{a highlight colour.}

\item{ideodat}{the data for ideograms.}
}
\value{
a grob with an ideogram
}
\description{
Ideogram renderer
}
