% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_constructors_continuous_colour.R,
%   R/scales_constructors_discrete_colour.R
\name{S4_colour_scales}
\alias{S4_colour_scales}
\alias{scale_colour_S4_continuous}
\alias{scale_fill_S4_continuous}
\alias{scale_colour_S4_gradient}
\alias{scale_fill_S4_gradient}
\alias{scale_colour_S4_viridis_c}
\alias{scale_fill_S4_viridis_c}
\alias{scale_colour_S4_hue}
\alias{scale_fill_S4_hue}
\alias{scale_colour_S4_discrete}
\alias{scale_fill_S4_discrete}
\title{S4 colour scales}
\description{
This collection of functions are analogues of ggplot2's
  colour/fill scales but use the S4 compatible scale system under the hood.
  Please refer to the ggplot2 documentation.
}
\section{Functions}{
\itemize{
\item \code{scale_colour_S4_continuous}: See
\code{\link[ggplot2]{scale_colour_continuous}}.

\item \code{scale_fill_S4_continuous}: See
\code{\link[ggplot2]{scale_colour_continuous}}.

\item \code{scale_colour_S4_gradient}: See
\code{\link[ggplot2]{scale_colour_gradient}}.

\item \code{scale_fill_S4_gradient}: See
\code{\link[ggplot2]{scale_colour_gradient}}.

\item \code{scale_colour_S4_viridis_c}: See
\code{\link[ggplot2]{scale_colour_viridis_c}}.

\item \code{scale_fill_S4_viridis_c}: See
\code{\link[ggplot2]{scale_colour_viridis_c}}.

\item \code{scale_colour_S4_hue}: See \code{\link[ggplot2]{scale_colour_hue}}

\item \code{scale_fill_S4_hue}: See \code{\link[ggplot2]{scale_colour_hue}}

\item \code{scale_colour_S4_discrete}: See \code{\link[ggplot2]{scale_colour_discrete}}

\item \code{scale_fill_S4_discrete}: See \code{\link[ggplot2]{scale_colour_discrete}}
}}

