% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_operations.R
\name{S4ExpandRange}
\alias{S4ExpandRange}
\alias{S4ExpandRange,NULL-method}
\alias{S4ExpandRange,numeric-method}
\alias{S4ExpandRange,Ranges-method}
\title{Range expansion}
\usage{
S4ExpandRange(limits, expand = expansion())

\S4method{S4ExpandRange}{`NULL`}(limits, expand = expansion())

\S4method{S4ExpandRange}{numeric}(limits, expand = expansion())

\S4method{S4ExpandRange}{Ranges}(limits, expand = expansion())
}
\arguments{
\item{limits}{An object indicating a range}

\item{expand}{A \code{numeric} vector of length 4, typically the result of a
call to \code{\link[ggplot2]{expansion}}.}
}
\value{
An object indicating a range
}
\description{
Expand a range with multiplicative and additive constants
}
\details{
The \code{expand} argument consist of 4 constants that code for the
  following:
  \enumerate{
   \item A multiplicative constant relative to the width of the \code{limits}
   argument to subtract from the start of the limits.
   \item An additive constant to subtract from the start of the limits.
   \item A multiplicative constant relative to the width of the \code{limits}
   argument to add to the end of the limits.
   \item An additive constant to add to the end of the limits.
  }

Integer-based classes such as \code{IRanges} and \code{GRanges} are rounded
to the nearest integer.
}
\examples{
# Regular numeric limits
S4ExpandRange(c(10, 20), expansion(0.5, 1))

# IntegerRanges
require(IRanges)
S4ExpandRange(IRanges::IRanges(10, 20), expansion(0.5, 1))

# GenomicRanges
require(GenomicRanges)
S4ExpandRange(GenomicRanges::GRanges(c("chr1", "chr2"), c("10-20", "20-30")),
              expansion(0.5, 1))
}
