% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-format.R
\name{S4LabelFormat}
\alias{S4LabelFormat}
\title{Format labels for S4 classes}
\usage{
S4LabelFormat(x, type = "major", ...)
}
\arguments{
\item{x}{A vector to format}

\item{type}{Either \code{"major"} or \code{"minor"} to determine how to
format the labels (default: \code{"major"}).}

\item{...}{Optional arguments passed to downstream functions (not
implemented).}
}
\value{
A \code{character} vector of the same length as \code{x} with labels
}
\description{
Attempts to figure out appropriate labels for S4 classes.
}
\details{
The \code{type} argument is currently only implemented for the
  \linkS4class{GenomicRanges} class.
}
\examples{
# Regular atomic vectors
S4LabelFormat(1:10)
S4LabelFormat(LETTERS[1:5])

# GenomicRanges major labels are seqnames
S4LabelFormat(GPos("chr1", 1:10))

# GenomicRanges minor labels are positions formatted as basepairs
S4LabelFormat(GPos("chr1", 1:10), type = "minor")
}
