% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_operations.R
\name{S4Range}
\alias{S4Range}
\alias{S4Range,knownContinuous-method}
\alias{S4Range,Rle-method}
\alias{S4Range,Ranges-method}
\alias{S4Range,GenomicRanges-method}
\alias{S4Range,WoodenHorse-method}
\title{Range of Values}
\usage{
S4Range(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")

\S4method{S4Range}{knownContinuous}(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")

\S4method{S4Range}{Rle}(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")

\S4method{S4Range}{Ranges}(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")

\S4method{S4Range}{GenomicRanges}(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")

\S4method{S4Range}{WoodenHorse}(x, ..., na.rm = FALSE, finite = FALSE, aes = "z")
}
\arguments{
\item{x}{An object to determine the range of.}

\item{...}{Optionally, more objects of the same class as \code{x} to include
in the range calculation.}

\item{na.rm}{\code{logical} of length 1: omit \code{NA} values prior to
calculating ranges?}

\item{finite}{\code{logical} of length 1: omit non-finite elements prior to
calculating ranges?}

\item{aes}{An aesthetic for which to evaluate the function.}
}
\value{
A vector indicating lower- and upper-bounds of the input.
}
\description{
Returns a vector containing the minimum and maximum of all the given
arguments.

Defaults to the \code{\link[base]{range}} function.
}
\examples{
# For regular numeric vectors
S4Range(c(1:3, -1))

# GenomicRanges returns same type
require(GenomicRanges)
S4Range(GRanges(c("chr1:100-200", "chr2:200-300", "chr2:500-600")))
}
