% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_operations.R
\name{S4ZeroRange}
\alias{S4ZeroRange}
\alias{S4ZeroRange,numeric-method}
\alias{S4ZeroRange,Ranges-method}
\title{Check zero rang}
\usage{
S4ZeroRange(x, tol = 1000 * .Machine$double.eps)

\S4method{S4ZeroRange}{numeric}(x, tol = 1000 * .Machine$double.eps)

\S4method{S4ZeroRange}{Ranges}(x, tol = 1000 * .Machine$double.eps)
}
\arguments{
\item{x}{An object representing a range}

\item{tol}{A value specifying the tolerance.}
}
\value{
A \code{logical} of length 1: \code{TRUE} if the relative differences
  of the range are not distinguishable from 0.
}
\description{
Check if range of limits is close to zero
}
\details{
Defaults to the \code{\link[scales]{zero_range}} function.
}
\examples{
S4ZeroRange(c(10, 10)) # TRUE
S4ZeroRange(c(10, 11)) # FALSE
S4ZeroRange(c(NA, 10)) # NA
}
