% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_constructors_misc.R
\name{scale_exontype_discrete}
\alias{scale_exontype_discrete}
\title{Scales for exon types.}
\usage{
scale_exontype_discrete(
  ...,
  palette = NULL,
  aesthetics = "exontype",
  guide = "none"
)
}
\arguments{
\item{...}{Arguments passed on to the S4 discrete scale}

\item{palette}{A palette function that returns a named numerical vector that
maps exon type names to widths.}

\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{guide}{A function used to create a guide or its name.}
}
\value{
A \code{ScaleS4} object.
}
\description{
Displaying gene models take the convention that coding sequences
  (CDSs) are drawn thicker than untranslated regions (UTRs). This scale type
  maps names of exontypes to thickness values.
}
\details{
This scale is typically used in conjunction with
  \code{\link[ggnomics]{geom_genemodel}}.

  By default, the following
  characters are recognised as CDSs: \code{"CDS"}, \code{"cds"}, and are
  drawn at 1x thickness. The following characters are recognised as UTRs:
  \code{"UTR"}, \code{"utr"}, \code{"three_prime_UTR"},
  \code{"three_prime_utr"}, \code{"five_prime_UTR"}, \code{"five_prime_utr"},
  and are drawn at 0.5x thickness.
}
\examples{
NULL
}
