% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodenHorse-class.R
\name{setNA}
\alias{setNA}
\title{set NAs}
\usage{
setNA(x, i)
}
\arguments{
\item{x}{An object to set NAs on}

\item{i}{A logical of \code{length(x)}, indicating which positions to set NAs.}
}
\value{

}
\description{
Not all S4 Vectors subclasses support having NAs.
}
\examples{
NULL
}
