% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideograms.R
\name{setup_cytobands}
\alias{setup_cytobands}
\title{Setting up cytogenetic bands}
\usage{
setup_cytobands(bands, colourmap, gpars = gpar())
}
\arguments{
\item{bands}{data.frame with the following columns in order:
\describe{\item{\code{character}  chrom}{Chromosome} \item{\code{numeric}
chromStart}{Start position in chromosome sequence} \item{\code{numeric}
chromEnd}{End position in chromosome sequence} \item{\code{character}
name}{Name of cytogenetic band} \item{\code{factor}  gieStain}{Giemsa stain
results}}}

\item{colourmap}{named vector of colours with names matching factor levels in
\code{bands$gieStain}.}

\item{gpars}{\code{gpar} object created by \code{\link[grid]{gpar}}}
}
\value{
None
}
\description{
Sets up the necessary \code{ggproto} objects and methods to plot ideograms
next to plot panels.
}
\details{
\code{bands} is assumed to follow the cytobands file format, but does
 not need matching column names. \code{bands$name} is not used but cannot be
 \code{NULL}. Special behaviour is applied to the "acen" factor level in
 \code{bands$gieStain}, as these are assumed to be the centromeres.

 Cytobands data can be downloaded from the Table Browser of the
 \href{https://genome.ucsc.edu/cgi-bin/hgTables}{UCSC Genome Browser} by
 setting the 'Group' to 'Mapping and Sequencing' and 'Track' to 'Chromosome
 Bands (Ideogram)' for the available organisms.  Alternatively, the
 \pkg{AnnotationHub} Bioconductor package can be used to retrieve cytoband
 information via an R interface.

 An appropriate \code{colourmap} vector for default Giemsa stain results can
 be generated by calling
 \code{\link[biovizBase:getBioColor]{getBioColor}}\code{("CYTOBANDS")} from
 the \pkg{biovizBase} Bioconductor Package.
}
\examples{
cytobands <- example_cytobands()
colourmap <- example_cytoband_colours()
setup_cytobands(cytobands, colourmap)

}
\seealso{
\href{https://genome.ucsc.edu/}{UCSC Genome Browser}
 \code{\link[biovizBase]{getBioColor}} \pkg{AnnotationHub}
 \code{\link[grid]{gpar}}
}
