% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ideograms.R
\name{setup_ideograms}
\alias{setup_ideograms}
\alias{setup_ideograms,data.frame,ANY-method}
\alias{setup_ideograms,character,ANY-method}
\alias{setup_ideograms,missing_OR_NULL,character-method}
\title{Setup ideograms}
\usage{
setup_ideograms(
  data = NULL,
  genome = NULL,
  colourmap = default_ideogram_colours()
)
}
\arguments{
\item{data}{One of the following:\itemize{
  \item{ A \code{data.frame} with the following 6 columns: \describe{
    \item{chrom}{A \code{character} column with chromosome names.}
    \item{chromStart}{An \code{integer} vector with band start positions.}
    \item{chromEnd}{An \code{integer} vector with band end positions.}
    \item{name}{A \code{character} with names for cytogenetic bands.}
    \item{gieStain}{A \code{character} with Giemsa stain results. Recognised
    values are: \code{"gneg"}, \code{"gpos"}, \code{"gpos50"},
    \code{"gpos75"}, \code{"gpos25"}, \code{"gpos100"}, \code{"acen"},
    \code{"gvar"}, and \code{"stalk"}.}}}
  \item{A \code{character} with the filename from which the data are to be
    read from.}
  \item{Can be \code{missing} if the \code{genome} argument is set. See
  details.}
}}

\item{genome}{A \code{character} with the genome name, e.g. \code{"mm10"}.}

\item{colourmap}{A named \code{character} vector with colours where names
correspond to the \code{gieStain} column.}
}
\value{
Returns no value, but caches the ideograms under the \code{genome}
  identifier in a temporary cache.
}
\description{
Sets up a local cache of ideogram data for a genome.
}
\details{
If the \code{data} argument is missing but the \code{genome}
  argument is either \code{"hg19"}, \code{"hg38"}, \code{"mm9"} or
  \code{"mm10"}, BiocFileCache is used to cache the \code{cytoBand.txt}
  file from UCSC.
}
\section{Methods (by class)}{
\itemize{
\item \code{data = data.frame,genome = ANY}: Uses \code{data.frame} to generate ideograms

\item \code{data = character,genome = ANY}: Uses \code{character} as file path.

\item \code{data = missing_OR_NULL,genome = character}: Attempts to lookup genome in UCSC tables and
caches this.
}}

\examples{
setup_ideograms(genome = "hg38")
}
\seealso{
This is to be used in conjunction with ideogram axes:
  \code{\link[ggnomics]{guide_ideogram_axis}()}
}
