% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_bounds.R
\name{squish_infiniteThis}
\alias{squish_infiniteThis}
\alias{squish_infiniteThis,numeric,numeric_OR_missing-method}
\alias{squish_infiniteThis,Rle,numeric_OR_missing-method}
\alias{squish_infiniteThis,WoodenHorse,numeric_OR_missing-method}
\title{Squish infinite values to range}
\usage{
squish_infiniteThis(x, range = NULL, aes = "z")

\S4method{squish_infiniteThis}{numeric,numeric_OR_missing}(x, range = c(0, 1), aes = "z")

\S4method{squish_infiniteThis}{Rle,numeric_OR_missing}(x, range = c(0, 1), aes = "z")

\S4method{squish_infiniteThis}{WoodenHorse,numeric_OR_missing}(x, range = c(0, 1), aes = "z")
}
\arguments{
\item{x}{A vector of values to manipulate.}

\item{range}{A range representing object.}

\item{aes}{An aesthetic for which to evaluate the function.}
}
\value{
The \code{x} argument, but with infinite values squished to the
  range.
}
\description{
This function follows the \code{\link[scales]{squish_infinite}}
  function, but allows extentions that work with other objects than numeric
  vectors.
}
\examples{
# For regular numeric vectors
squish_infiniteThis(c(-Inf, -1, 0, 1, 2, Inf))

# For Rle-class
squish_infiniteThis(Rle(c(Inf, 2, 3, -Inf), 1:4), c(2, 3))
}
