% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_bounds.R
\name{squishThis}
\alias{squishThis}
\alias{squishThis,WoodenHorse,ANY-method}
\alias{squishThis,numeric,numeric_OR_missing-method}
\alias{squishThis,Rle,numeric_OR_missing-method}
\alias{squishThis,IntegerRanges,numeric_OR_missing-method}
\alias{squishThis,ANYGenomic,GRanges-method}
\title{Squish values into range}
\usage{
squishThis(x, range = NULL, only.finite = TRUE, aes = "z")

\S4method{squishThis}{WoodenHorse,ANY}(x, range = c(0, 1), only.finite = TRUE, aes = "z")

\S4method{squishThis}{numeric,numeric_OR_missing}(x, range = c(0, 1), only.finite = TRUE, aes = "z")

\S4method{squishThis}{Rle,numeric_OR_missing}(x, range = c(0, 1), only.finite = TRUE, aes = "z")

\S4method{squishThis}{IntegerRanges,numeric_OR_missing}(x, range = c(0, 1), only.finite = TRUE, aes = "z")

\S4method{squishThis}{ANYGenomic,GRanges}(x, range = GRanges(NA_character_, IRanges(0, 1)))
}
\arguments{
\item{x}{A vector of values to manipulate.}

\item{range}{A range representing object.}

\item{only.finite}{If \code{TRUE} (the default), will only modify finite
values.}

\item{aes}{An aesthetic for which to evaluate the function.}
}
\value{
The \code{x} argument, but with out-of-bounds values squished to the
  range.
}
\description{
This function follows the \code{\link[scales]{squish}} function, but allows
extentions that work with objects other than numeric vectors.
}
\examples{
# For regular numeric vectors
squishThis(c(-1, 0.5, 1, 2, NA))

# For Rle-class
squishThis(Rle(1:4, 4:1), c(2,3))
}
