% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_coverage.R
\name{stat_coverage}
\alias{stat_coverage}
\title{Coverage}
\usage{
stat_coverage(
  mapping = NULL,
  data = NULL,
  geom = "area",
  position = "identity",
  ...,
  trim = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{Use to override the default connection between
\code{geom_density} and \code{stat_density}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{trim}{If \code{FALSE}, the default, each density is computed on the
full range of the data. If \code{TRUE}, each density is computed over the
range of that group: this typically means the estimated x values will
not line-up, and hence you won't be able to stack density values.
This parameter only matters if you are displaying multiple densities in
one plot or if you are manually adjusting the scale limits.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetics mapping. In
the rare event that this fails, it can be given explicitly by setting
\code{orientation} to either \code{"x"} or \code{"y"}. See the
\emph{Orientation} section of
\code{\link[ggplot2]{geom_density}()}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Calculates how often a space is covered by a set of ranges.
}
\details{
For \linkS4class{IRanges} and \linkS4class{GRanges} classes, makes
use of the \code{\link[IRanges]{coverage}} function. When data is
\code{numeric}, be sure to also set the \code{xend} or \code{yend} aesthetic.
}
\note{
The \code{Ranges} classes as \code{x|y} are considered closed-interval,
whereas \code{x|y} and \code{xend|yend} with numeric data are considered open
interval.
}
\section{Aesthetics}{
 \code{stat_coverage} understands the following
aesthetics (required aesthetics are in bold, optional in italic).
\itemize{
 \item \strong{x}
 \item \strong{y}
 \item \emph{xend} (if data is numeric)
 \item \emph{yend} (if data is numeric)
 \item alpha
 \item colour
 \item fill
 \item group
 \item linetype
 \item size
 \item weight
}
}

\section{Computed variables}{

\describe{
  \item{coverage}{coverage}
  \item{scaled}{coverage scaled to a maximum of 1}
  \item{density}{coverage integrated to 1}
}
}

\examples{
# Computing coverage on numeric data
df <- data.frame(min = c(1, 6, 11),
                 max = c(9, 13, 15))
ggplot(df, aes(x = min, xend = max)) +
  stat_coverage()

# Computing coverage on Ranges data
require(GenomicRanges)

df <- DataFrame(x = GRanges(c("chr1:100-200", "chr1:140-260",
                              "chr2:50-100")))
ggplot(df, aes(x)) +
  stat_coverage()
}
