% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_genemodel.R
\name{style_intron_arrowline}
\alias{style_intron_arrowline}
\alias{style_intron_plainline}
\alias{style_intron_chevron}
\title{Intron styles}
\usage{
style_intron_arrowline(
  data,
  linejoin = "mitre",
  arrow.template = NULL,
  arrow.freq = unit(4, "mm")
)

style_intron_plainline(data, linejoin)

style_intron_chevron(data, linejoin = "bevel", height = 1)
}
\description{
Curious little monkey, aren't we? These internal helper
  functions help \code{geom_genemodel} style its introns. I can't see why a
  user would need these as they only make sense in the context of
  constructing grobs. Alas, here we are: me trying to hide my internal
  functions and you reading poorly written documentation for them anyway.
}
\seealso{
\code{\link[ggnomics]{geom_genemodel}}
}
\keyword{internal}
