% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_poly.R
\name{to_poly}
\alias{to_poly}
\alias{to_poly,BigWigFile,GenomicRanges-method}
\alias{to_poly,RleList,GenomicRanges-method}
\alias{to_poly,RleList,missing-method}
\alias{to_poly,Rle,missing-method}
\alias{to_poly,rle,missing-method}
\title{Covert an object to polygon data.frame}
\usage{
to_poly(object, selection)

\S4method{to_poly}{BigWigFile,GenomicRanges}(object, selection)

\S4method{to_poly}{RleList,GenomicRanges}(object, selection)

\S4method{to_poly}{RleList,missing}(object)

\S4method{to_poly}{Rle,missing}(object)

\S4method{to_poly}{rle,missing}(object)
}
\arguments{
\item{object}{An object to convert to a polygon.}

\item{selection}{A subset indication of the object.}
}
\value{
A data.frame with a parameterised polygon.
}
\description{
Covert an object to polygon data.frame
}
\details{
At the moment, this function supports converting objects of the
  following classes: \code{\link[rtracklayer]{BigWigFile}},
  \code{\link[IRanges]{RleList}}, \code{\link[S4Vectors]{Rle}} and
  \code{\link[base]{rle}} to polygons. Note that the \code{BigWigFile}
  requires a \code{\link[GenomicRanges]{GRanges}} selection argument.

  For list-like objects, group IDs are automatically assigned. For objects
  with a selection containing genomic coordinates, sequence names are
  included if there are multiple.
}
\examples{
x <- rle(rep(1:5, each = 5))
to_poly(x)
}
