% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_track.R
\name{tracklayer}
\alias{tracklayer}
\title{Create a layer as track}
\usage{
tracklayer(
  data = NULL,
  mapping = NULL,
  geom = "point",
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  check.aes = TRUE,
  check.param = TRUE,
  show.legend = NA,
  key_glyph = NULL,
  track_y = "Default",
  track_x = "Default",
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{geom}{The geometric object to use display the data}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{check.aes}{If \code{TRUE}, the default, will check that
supplied parameters and aesthetics are understood by the \code{geom} or
\code{stat}. Use \code{FALSE} to suppress the checks.}

\item{check.param}{If \code{TRUE}, the default, will check that
supplied parameters and aesthetics are understood by the \code{geom} or
\code{stat}. Use \code{FALSE} to suppress the checks.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{key_glyph}{A legend key drawing function or a string providing the
function name minus the \code{draw_key_} prefix. See \link[ggplot2]{draw_key} for details.}

\item{track_y}{A \code{character} of length 1 setting the name of a facet row
for this layer. Alternatively, \code{track_y} may also be defined in the
aesthetics mapping.}

\item{track_x}{A \code{character} of length 1 setting the name of a facet
column for this layer. Alternatively, \code{track_x} may also be defined in
the aesthetics mapping.}

\item{...}{Other arguments passed on to the geom, stat or position
components. These are often aesthetics, used to set an aesthetic to a fixed
value, like \code{colour = "red"} or \code{size = 3}. They may also be
parameters to the paired geom/stat.}
}
\description{
\code{tracklayer} moves the responsibilities of setting facet specifications
from the facet function to the layer, which allows defining facets at the
layer level.
}
\details{
There are two behaviours of \code{tracklayer} layers which are
  different from the regular \code{layer} function.

  First, before facet are evaluated, \code{tracklayer} attaches the
  \code{"track_y"} and \code{"track_x"} columns to the data, allowing these
  to be found by the facet functions. These can either be provided as
  arguments to the layer or within aesthetic mappings; the latter will be
  prioritised over the former.

  Secondly, when a \code{tracklayer} is added to a plot, it will check if the
  plot already has non-default facets. If the plot has the default facet
  (i.e. as produced by \code{facet_null()}), this facet is replaced by
  \code{facet_track} that only takes \code{track_x} and \code{track_y} as
  facets.
}
\examples{
# Tracked layer as argument
ggplot() +
  tracklayer(mapping = aes(wt, drat),
             data = mtcars, track_y = "mtcars") +
  tracklayer(mapping = aes(Sepal.Width, Sepal.Length, colour = Species),
             data = iris, track_y = "iris")

# Tracked layer through aesthetics mapping
ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  tracklayer(mapping = aes(colour = Species, track_y = Species))
  
# 'track_y' and 'track_x' can also be used with conventional facets
ggplot(iris, aes(Sepal.Width, Sepal.Length)) +
  tracklayer(mapping = aes(colour = Species, track_x = Species)) +
  facet_wrap(~ track_x)
}
\seealso{
\code{\link[ggnomics]{facet_track}} for the sister function in
  laying out tracked facets. \code{\link[ggplot2]{layer}} for the
  conventional use of layers.
}
