% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_cytobands}
\alias{example_cytobands}
\title{Example cytobands data}
\usage{
example_cytobands()
}
\value{
A \code{data.frame} containing cytoband data for the first two
  chromosomes of the hg38 human genome.
}
\description{
Generates some toy cytoband data to plot ideograms with.
}
\examples{
cyto <- example_cytobands()
}
\seealso{
\code{\link[ggnomics]{example_cytoband_colours}}
}
