% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_ideogrid.R
\name{facet_ideogrid}
\alias{facet_ideogrid}
\title{Lay out panels in a grid with ideograms}
\usage{
facet_ideogrid(
  rows = NULL,
  cols = NULL,
  scales = "fixed",
  space = "fixed",
  shrink = TRUE,
  labeller = "label_value",
  as.table = TRUE,
  switch = NULL,
  drop = TRUE,
  margins = FALSE,
  facets = NULL,
  ideo.size = unit(0.1, "null"),
  high.col = NA
)
}
\arguments{
\item{rows}{A set of variables or expressions quoted by
\code{\link[ggplot2:vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{cols}{A set of variables or expressions quoted by
\code{\link[ggplot2:vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with formulae of the type \code{~cyl + am}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. See
\code{\link[ggplot2:label_value]{label_value()}} for more details and pointers to other
options.}

\item{as.table}{If \code{TRUE}, the default, the facets are laid out like
a table with highest values at the bottom-right. If \code{FALSE}, the
facets are laid out like a plot with the highest value at the top-right.}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{drop}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}

\item{margins}{Either a logical value or a character
vector. Margins are additional facets which contain all the data
for each of the possible values of the faceting variables. If
\code{FALSE}, no additional facets are included (the
default). If \code{TRUE}, margins are included for all faceting
variables. If specified as a character vector, it is the names of
variables for which margins are to be created.}

\item{facets}{This argument is soft-deprecated, please use \code{rows}
and \code{cols} instead.}

\item{ideo.size}{\code{\link[grid]{unit}} value object of class \code{"unit"}
specifying ideogram size.}

\item{high.col}{\code{NA} (default) or colour value of ideogram highlights
(see details). If \code{NA}, no highlighting will occur.}
}
\description{
\code{facet_ideogrid()} mimicks behaviour of
\code{\link[ggplot2]{facet_grid}} but adds ideograms. The facetting variables
are checked for chromosome names and if these are found, then ideograms are
plotted in between the strips and panels.
}
\details{
In absence of chromosome names in variables will result in default
  \code{facet_grid()} behaviour. A mix of chromosome names and other
  variables results in whitespace between the panels and other variable
  strips.

  By default, this function will not highlight the panel axis range. To
  highlight the plotted region on the ideogram, set \code{high.col} to a
  colour. If highlights occur, region to be highlighted will be computed from
  the relevant axis limits of the corresponding panels.

  Colours can be specified in one of the forms returned by
  \code{\link[grDevices]{rgb}}, as name (see \code{\link[grDevices]{colors}})
  or as non-negative integer index into the current
  \code{\link[grDevices]{palette}} (with zero being taken as transparent).
}
\note{
\code{\link[ggnomics]{setup_cytobands}} needs to be run to populate
  ideogram data before calling \code{facet_ideogrid}.
}
\examples{
setup_cytobands(example_cytobands(),
                example_cytoband_colours())

df <- cbind(mpg, chr = sample(c("chr1", "chr2"), nrow(mpg), replace = TRUE))
ggplot(df, aes(displ, cty)) +
  geom_point() +
  facet_ideogrid(~ chr)
}
\seealso{
\code{\link[ggplot2]{facet_grid}} \code{\link[grid]{unit}}
}
