% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_nested.R
\docType{data}
\name{facet_nested}
\alias{facet_nested}
\alias{FacetNested}
\title{Layout panels in a grid with nested strips}
\usage{
facet_nested(rows = NULL, cols = NULL, scales = "fixed",
  space = "fixed", shrink = TRUE, labeller = "label_value",
  as.table = TRUE, switch = NULL, drop = TRUE, margins = FALSE,
  facets = NULL, nest_line = FALSE, resect = unit(0, "mm"))
}
\arguments{
\item{rows, cols, scales, space, shrink, labeller, as.table, switch, drop, margins, facets}{see \code{\link[ggplot2]{facet_grid}}}

\item{nest_line}{a \code{logical} vector of length 1, indicating whether to
draw a nesting line to indicate the nesting of variables.}

\item{resect}{a \code{unit} vector of length 1, indicating how much the
nesting line should be shortened.}
}
\description{
\code{facet_nest()} form a matrix of panels defined by row and
  column faceting variables and nests grouped facets.
}
\details{
Unlike \code{facet_grid()}, this function only automatically expands
  missing variables when they have no variables in that direction, to allow
  for unnested variables. It still requires at least one layer to have all
  faceting variables.

  Hierarchies are inferred from the order of variables supplied to
  \code{rows} or \code{cols}. The first variable is interpreted to be the
  outermost variable, while the last variable is interpreted to be the
  innermost variable. They display order is always such that the outermost
  variable is placed the furthest away from the panels.

  Strips are automatically grouped when they span a nested variable.
}
\examples{
df <- iris
df$nester <- ifelse(df$Species == "setosa",
                    "Short Leaves",
                    "Long Leaves")

ggplot(df, aes(Sepal.Length, Petal.Length)) +
  geom_point() +
  facet_nested(~ nester + Species)

}
\seealso{
See \code{\link[ggplot2]{facet_grid}} for descriptions of the
  original arguments. See \code{\link[grid]{unit}} for the construction of a
  \code{unit} vector.
}
\keyword{datasets}
