% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hictriangle.R
\name{geom_hictriangle}
\alias{geom_hictriangle}
\title{Hi-C Triangle}
\usage{
geom_hictriangle(exp, ranges, stat = "identity", position = "identity",
  ..., na.rm = FALSE, show.legend = NA)
}
\arguments{
\item{exp}{A \pkg{GENOVA} experiment}

\item{ranges}{A GRanges object. Alternatively, a \code{list} of at least
length 3 with the following elements: \describe{\item{\code{character}}{
vector of chromosome names}\item{\code{integer}}{ vector of start
positions}\item{\code{integer}}{ vector of end positions}}}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\description{
The Hi-C triangle geom is used to portray a rotated Hi-C matrix such that the
diagonal runs along the x-axis. It can be convenient for displaying details
located in a region of interest and proximal to the diagonal.
}
\details{
The triangle is calculated such that it is rotated by 45 degrees
  clockwise and rescaled such that the x-axis is equivalent to the original
  region and the y-axis corresponds to distance from the diagonal.
  \code{geom_hictriangle} invokes these steps at the layer level by
  specifying how a raster-like Hi-C matrix should be setup as a triangle.

  A recommendation for this geom is to use a \code{coord_fixed(ratio = 0.5)}
  to get neat 45 degree angles.
}
\examples{
require(GenomicRanges)
exp <- example_HiC()
gr <- GRanges("chr1", IRanges(20e6, 100e6))
ggplot() +
  geom_hictriangle(exp, gr)
}
\seealso{
\code{\link[GENOVA]{construct.experiment}} on how to construct a
  \pkg{GENOVA} experiment object. \code{\link[GenomicRanges]{GRanges}} on how to
  make or subset GRanges objects.
}
