% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnomics-package.R
\docType{package}
\name{ggnomics-package}
\alias{ggnomics}
\alias{ggnomics-package}
\title{ggnomics: ggnomics}
\description{
This package contains some extensions to ggplot2 and convenience functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teunbrand/ggnomics}
}

}
\keyword{internal}
