% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewscales.R, R/coord_class_cartesian.R,
%   R/ggnomics_extensions.R, R/position_disjoint_ranges.R, R/range_ggproto.R,
%   R/scales_class_S4.R, R/scales_class_genomic.R
\docType{data}
\name{ViewScaleS4}
\alias{ViewScaleS4}
\alias{ViewScaleS4Secondary}
\alias{CoordS4}
\alias{ggnomics_extensions}
\alias{PositionDisjointRanges}
\alias{RangeS4}
\alias{RangeS4Continuous}
\alias{RangeS4Discrete}
\alias{ScaleS4}
\alias{ScaleS4Continuous}
\alias{ScaleS4Discrete}
\alias{ScaleS4ContinuousPosition}
\alias{ScaleGenomic}
\title{ggnomics extensions to ggplot2}
\description{
ggnomics relies on the extension mechanism of ggplot2 through
  ggproto class objects, which allows cross-package inheritance of objects
  such as geoms, stats, facets, scales and coordinate systems. These objects
  can be ignored by users for the purpose of making plots, since interacting
  with these objects is preferred through various geom_\*, stat_\*, facet_\*,
  coord_\* and scale_\* functions.
  
  Note that the below says \strong{Functions}, but actually describes ggproto
  classes.
}
\section{Functions}{
\itemize{
\item \code{ViewScaleS4}: An child to ggplot's ViewScale ggproto that
has a minor label getter. Note: this class is not exported.

\item \code{ViewScaleS4Secondary}: A child to ViewScaleS4 that has static break
information. Note: this class is not exported.

\item \code{CoordS4}: See \code{\link[ggnomics]{coord_S4}}

\item \code{PositionDisjointRanges}: See
\code{\link[ggnomics]{position_disjoint_ranges}}.

\item \code{RangeS4}: Identical to ggplot's Range ggproto. Note:
this class and children are not exported and are for internal use.

\item \code{RangeS4Continuous}: A child to RangeS4 with S4 compatible train
method.

\item \code{RangeS4Discrete}: A child to RangeS4 with S4 compatible train
method.

\item \code{ScaleS4}: A child to ggplot's Scale ggproto. Note that
this class is not exported but children are.

\item \code{ScaleS4Continuous}: A child to ScaleS4 and sibling to ggplot's
ScaleContinuous.

\item \code{ScaleS4Discrete}: A child to ScaleS4 and sibling to ggplot's
ScaleDiscrete

\item \code{ScaleS4ContinuousPosition}: A child to ScaleS4Continuous and sibling to
ggplot's ScaleContinuousPosition. See
\code{\link[ggnomics]{scale_S4_continuous}}.

\item \code{ScaleGenomic}: A child to ScaleS4ContinuousPosition for
genomic coordinates. See \code{\link[ggnomics]{scale_genomic}}.
}}

\seealso{
\link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
