% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek_soldier.R
\name{TheFallOfTroy}
\alias{TheFallOfTroy}
\alias{GreekSoldier}
\alias{Nightfall}
\alias{HelenOfTroy}
\alias{GreekSoldier,ANY-method}
\alias{GreekSoldier,Vector-method}
\alias{GreekSoldier,AtomicList-method}
\alias{GreekSoldier,ANYGenomic-method}
\alias{GreekSoldier,matrix-method}
\alias{Nightfall,ANY-method}
\alias{Nightfall,BeechHorse-method}
\alias{Nightfall,OakHorse-method}
\title{Hiding S4 objects as S3 vectors}
\usage{
GreekSoldier(x)

Nightfall(x, na.rm = FALSE)

\S4method{Nightfall}{OakHorse}(x, na.rm = FALSE)

HelenOfTroy(x)
}
\arguments{
\item{x}{For \code{GreekSoldier()}, an S4 Vector class object.\cr For
\code{Nightfall()} and \code{HelenOfTroy()}, an S3 \code{WoodenHorse} class
object.}

\item{na.rm}{For \code{Nightfall()}, should the result be returned for
non-NA entries?}
}
\value{
For \code{Nightfall()}, a \code{Vector} object. For
  \code{GreekSoldier()}, a \code{WoodenHorse} object. For
  \code{HelenaOfTroy()}, a \code{character},
}
\description{
There are some limitations to getting S4 objects past the
  internals of the ggplot2 system. One of these is a hard check with
  \code{rlang::is_vector} on data.frame columns, which will throw an error
  when the column is an S4 object. To work around these limitations, we
  disguise S4 Vectors classes as S3 vectors from the vctrs package.
}
\details{
Calling \code{GreekSoldier()} on an S4 Vector object will generate
  an object of the class \code{WoodenHorse} of the same length as the input,
  where the original S4 Vector is inside the \code{WoodenHorse} object as an
  attribute.

  Encoding the \code{WoodenHorse} object as a \code{vctrs_vctr} object allows
  us to pass the \code{rlang::is_vector} check and gives us nice options to
  preserve attributes over subsetting and concatenation operations.
}
\section{Functions}{
\itemize{
\item \code{GreekSoldier,ANY-method}: By default, don't disguise object as
\code{WoodenHorse} and return the input.

\item \code{GreekSoldier,Vector-method}: If an object inherits from \linkS4class{Vector},
disguise object as \code{WoodenHorse}, and put the input as an attribute of
the vector.

\item \code{Nightfall,ANY-method}: If the object is not a \code{WoodenHorse} return
the input unaltered.

\item \code{Nightfall,BeechHorse-method}: If the object is a \code{WoodenHorse} return the
\code{GreekSoldier} attribute within the \code{WoodenHorse}.

\item \code{Nightfall,OakHorse-method}: If the object is a \code{WoodenHorse} return the
\code{GreekSoldier} attribute within the \code{WoodenHorse}.
}}

\examples{
# Making an S4 object
Anticlus <- Rle(1:10, 10:1)

# Rle object does not pass is_vector test
rlang::is_vector(Anticlus)

# Disguising the object as S3 vector
VictoryTrophy <- GreekSoldier(Anticlus)
class(VictoryTrophy) # A WoodenHorse

# WoodenHorse passes is_vector test
rlang::is_vector(VictoryTrophy)

# Inspecting the class of WoodenHorse
HelenOfTroy(VictoryTrophy) # Rle

# Restoring WoodenHorse to S4 object
Nightfall(VictoryTrophy)
}
