% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodenHorse-class.R
\name{WoodenHorse-vctr}
\alias{WoodenHorse-vctr}
\alias{obj_print_header.WoodenHorse}
\alias{vec_ptype_full.WoodenHorse}
\alias{vec_ptype_abbr.WoodenHorse}
\alias{vec_ptype2.WoodenHorse}
\alias{vec_ptype2.WoodenHorse.default}
\alias{vec_ptype2.WoodenHorse.WoodenHorse}
\alias{vec_cast.WoodenHorse}
\alias{vec_cast.WoodenHorse.default}
\alias{vec_cast.WoodenHorse.WoodenHorse}
\alias{vec_math.WoodenHorse}
\alias{vec_arith.WoodenHorse}
\alias{vec_arith.WoodenHorse.default}
\alias{vec_arith.WoodenHorse.WoodenHorse}
\title{WoodenHorse vctr methods}
\description{
Here are the methods listed for the \code{WoodenHorse} class.
  Refer to the documentation in the \code{vctrs} package to learn about their
  use.
}
\details{
These are vctrs-methods that enable the \code{WoodenHorse} class,
  but should never be called directly by users. Consider these
  implementation details.
}
\section{Functions}{
\itemize{
\item \code{obj_print_header.WoodenHorse}: See \code{\link[vctrs]{obj_print_header}}

\item \code{vec_ptype_full.WoodenHorse}: See \code{\link[vctrs]{vec_ptype_full}}

\item \code{vec_ptype_abbr.WoodenHorse}: See \code{\link[vctrs]{vec_ptype_abbr}}

\item \code{vec_ptype2.WoodenHorse}: Generic for \code{\link[vctrs]{vec_ptype2}}

\item \code{vec_ptype2.WoodenHorse.default}: Default method for
\code{vec_ptype2.WoodenHorse}: there are no common types.

\item \code{vec_ptype2.WoodenHorse.WoodenHorse}: WoodenHorse method for
\code{vec_ptype2.WoodenHorse}: attempt to merge
\code{GreekSoldier} attributes.

\item \code{vec_cast.WoodenHorse}: Generic for \code{\link[vctrs]{vec_cast}}

\item \code{vec_cast.WoodenHorse.default}: Default method for \code{vec_cast.WoodenHorse}:
it should not be casted.

\item \code{vec_cast.WoodenHorse.WoodenHorse}: \code{WoodenHorse} method for
\code{vec_cast.WoodenHorse}: attempt to cast if the \code{GreekSoldier}
attribute is similar.

\item \code{vec_math.WoodenHorse}: See \code{\link[vctrs]{vec_math}}. Exceptions to
the math group generics for \code{WoodenHorse} class are \code{is.na()},
\code{is.finite()} and \code{is.infinite()}.

\item \code{vec_arith.WoodenHorse}: Generic for \code{\link[vctrs]{vec_arith}}

\item \code{vec_arith.WoodenHorse.default}: Default method for \code{vec_arith.WoodenHorse}:
try and apply arithmatic to the \code{GreekSoldier} attribute.

\item \code{vec_arith.WoodenHorse.WoodenHorse}: \code{WoodenHorse} method for
\code{vec_arith.WoodenHorse}: try and apply arithmatic between the
\code{GreekSoldier} attribute of the two vectors.
}}

\seealso{
See \code{\link[ggnomics]{TheFallOfTroy}} for the role that the
  \code{WoodenHorse} plays in the package.
}
